Run based on whether we've seen every day in the month.
Currently we're stuck waiting 'til the logs for the first day of the
next month are available before we start processing, since that's the
easiest way to tell that we've actually got every day for the
month. That annoys me, so here's a tweak that instead makes sure we've
got at least one logfile for every day of a given month and proceeds
if that's the case.